Добавила подгрузку изображений с сервера
