+++ js/api.js
@@ -1,29 +1,23 @@
-// export const elementsLoader = (onSuccess, onError) => () => fetch(
-//   'https://31.javascript.htmlacademy.pro/kekstagram/data',
-//   {
-//     method: 'GET',
-//     credentials: 'same-origin',
-//   },
-// )
-//   .then((response) => {
-//     if (response.ok) {
-//       return response.json();
-//     }
-//
-//     throw new Error(`${response.status} ${response.statusText}`);
-//   })
-//   .then((data) => {
-//     onSuccess(data);
-//   })
-//   .catch((err) => {
-//     onError(err);
-//   });
-const DATA_URL = 'https://31.javascript.htmlacademy.pro/kekstagram/data';
+export const BASE_URL = 'https://31.javascript.htmlacademy.pro/kekstagram';
 
-export const loadPhotos = async () => {
-  const response = await fetch(DATA_URL);
+export const Route = {
+  GET_DATA: '/data',
+  SEND_DATA: '/',
+}
+export const Method = {
+  GET: 'GET',
+  POST: 'POST',
+}
+export const ErrorText = {
+  [Method.GET]: 'Не удалось загрузить данные. Попробуйте еще раз',
+  [Method.POST]: 'Не удалось отправить данные формы',
+}
+
+export const getData = async (route, method, body = null) => {
+  const response = await fetch(`${BASE_URL}${route}`, { method, body});
   if (!response.ok) {
-    throw new Error(`Ошибка ${response.status}: ${response.statusText}`);
+    throw new Error(ErrorText[method]);
   }
   return response.json();
 };
+
