+++ js/main.js
@@ -1,6 +1,7 @@
-import './photos.js';
-import './open-photo.js';
+// import './photos.js';
+// import './open-photo.js';
 import './form.js';
 import './slider.js';
 import './api.js';
 
+import './picture-from-api.js';
