+++ js/picture-from-api.js
@@ -1,0 +1,182 @@
+import { photoTemplate, errorTemplate } from "./templates.js";
+import {getData, Method, Route, ErrorText, getData} from "./api.js";
+
+export const container = document.querySelector('.pictures');
+const photosTotal = [];
+
+const COMMENTS_ON_SHOW = 5;
+
+const openBigPicture = document.querySelector('.big-picture');
+const bigPicture = document.querySelector('.big-picture__img');
+const closeBigPicture = document.querySelector('.big-picture__cancel');
+const showMoreComments = openBigPicture.querySelector('.social__comments-loader');
+
+function changeEventListeners(action) {
+  const method = `${action}EventListener`;
+  document[method]('keydown', handleEscapeKey);
+  closeBigPicture[method]('click', closeModal);
+}
+
+let currentLoadMoreHandler = null;
+const loadAndRenderPhotos = async () => {
+  try {
+    const photosData = await getData(Route.GET_DATA, Method.GET);
+    console.log('Данные загружены:', photosData);
+
+    const photosListFragment = document.createDocumentFragment();
+
+    //Создаем миниатюры для каждой подгруженной с сервера фотки
+
+    photosData.forEach((picture) => {
+      const template = photoTemplate.cloneNode(true);
+      template.photoData = picture;
+      console.log('Pictures comments are: ', picture.comments);
+
+      const image = template.querySelector('.picture__img');
+      image.alt = picture.description;
+      image.src = picture.url;
+
+      template.querySelector('.picture__likes').textContent = picture.likes;
+      template.querySelector('.picture__comments').textContent = picture.comments.length;
+
+      photosListFragment.appendChild(template);
+      photosTotal.push(picture);
+
+      template.addEventListener('click', () => {
+        onClickOpenModal(template);
+      });
+    });
+
+    container.appendChild(photosListFragment);
+
+  } catch (error) {
+    showDataError(Method.GET);
+  }
+};
+
+function showDataError(method) {
+
+  const errorElement = errorTemplate.content.cloneNode(true);
+  errorElement.querySelector('h2').textContent = ErrorText[method];
+
+  document.body.appendChild(errorElement);
+
+  setTimeout(() => {
+    const errorSection = document.querySelector('.data-error');
+    if (errorSection) {
+      errorSection.remove();
+    }
+  }, 5000);
+}
+
+//Рендерим миниатюры фото и пытаемся открыть модалку
+document.addEventListener('DOMContentLoaded', loadAndRenderPhotos);
+
+
+function handleEscapeKey(evt) {
+  if (evt.key === 'Escape') {
+    console.log('ESC detected, closing modal');
+    closeModal();
+  }
+}
+
+function closeModal() {
+  console.log('closeModal called');
+  openBigPicture.classList.add('hidden');
+  document.querySelector('body').classList.remove('modal-open');
+
+  changeEventListeners('remove');
+
+  if (currentLoadMoreHandler) {
+    showMoreComments.removeEventListener('click', currentLoadMoreHandler);
+    currentLoadMoreHandler = null;
+  }
+}
+
+function renderComments(comments, commentList, fragment) {
+  comments.forEach((comment) => {
+    const template = commentList.querySelector('li').cloneNode(true);
+
+    const image = template.querySelector('.social__picture');
+    image.alt = comment.name;
+    image.src = comment.avatar;
+
+    const commentText = template.querySelector('.social__text');
+    commentText.textContent = comment.message;
+
+    fragment.appendChild(template);
+  });
+  return fragment;
+}
+function onClickOpenModal (thumbnail) {
+  if (currentLoadMoreHandler) {
+    showMoreComments.removeEventListener('click', currentLoadMoreHandler);
+    currentLoadMoreHandler = null;
+  }
+
+  document.removeEventListener('keydown', handleEscapeKey);
+  closeBigPicture.removeEventListener('click', closeModal);
+
+  const commentAmount = openBigPicture.querySelector('.social__comment-total-count');
+  const commentList = openBigPicture.querySelector('.social__comments');
+  const commentFragment = document.createDocumentFragment();
+
+  openBigPicture.classList.remove('hidden');
+  document.querySelector('body').classList.add('modal-open');
+
+  bigPicture.querySelector('img').src = thumbnail.querySelector('img').src;
+  bigPicture.querySelector('img').alt = thumbnail.querySelector('img').alt;
+
+  openBigPicture.querySelector('.likes-count').textContent = thumbnail.querySelector('.picture__likes').textContent;
+  openBigPicture.querySelector('.social__caption').textContent = thumbnail.querySelector('img').alt;
+
+  const thisPhotoComments = thumbnail.photoData.comments;
+  console.log(thisPhotoComments);
+
+  commentAmount.textContent = thisPhotoComments.length.toString();
+  console.log('Кол-во комментариев к фото: ',thisPhotoComments.length );// ОСТАНОВИЛАСЬ ЗДЕСЬ
+
+  if (thisPhotoComments.length <= 5) { //Показываем 5 комментов
+    openBigPicture.querySelector('.comments-loader').classList.add('hidden');
+    openBigPicture.querySelector('.social__comment-shown-count').textContent = commentAmount.textContent;
+
+    renderComments(thisPhotoComments, commentList, commentFragment);
+    commentList.replaceChildren(commentFragment);
+
+  } else {
+    openBigPicture.querySelector('.comments-loader').classList.remove('hidden');
+
+    let shownCommentsAmount = COMMENTS_ON_SHOW;
+
+    openBigPicture.querySelector('.social__comment-shown-count').textContent = shownCommentsAmount;
+
+    const restComments = [...thisPhotoComments];
+    let showingComments = restComments.splice(0, COMMENTS_ON_SHOW);
+
+    renderComments(showingComments, commentList, commentFragment);
+
+    commentList.replaceChildren(commentFragment);
+
+    currentLoadMoreHandler = () => {
+      if (restComments.length > 0) {
+        showingComments = restComments.splice(0, COMMENTS_ON_SHOW);
+        shownCommentsAmount += showingComments.length;
+
+        renderComments(showingComments, commentList, commentFragment);
+
+        commentList.append(commentFragment);
+        openBigPicture.querySelector('.social__comment-shown-count').textContent = shownCommentsAmount;
+      }
+
+      if (restComments.length === 0) {
+        showMoreComments.classList.add('hidden');
+      }
+    };
+
+    showMoreComments.addEventListener('click', currentLoadMoreHandler);
+  }
+  console.log('Adding event listeners'); // ← ДОБАВЬТЕ
+  changeEventListeners('add');
+
+  console.log('Event listeners added'); // ← ДОБАВЬТЕ
+}
