+++ js/templates.js
@@ -1,2 +1,5 @@
 export const photoTemplate = document.querySelector('#picture').content.querySelector('.picture');
+export const errorTemplate = document.querySelector('#data-error');
+export const sendTemplate = document.querySelector('#success');
+export const sendErrorTemplate = document.querySelector('#error');
